--- src/app/(root)/lifestyle/lifestyleHeader.tsx
@@ -1,10 +1,0 @@
-export default function LifestyleHeader() {
-  return (
-    <div className="flex flex-col items-center py-10 px-4">
-      <p className="text-4xl font-bold text-gray-800">Lifestyle Collection</p>
-      <p className="text-gray-600 mt-2">
-        Please choose your lifestyle from this collection
-      </p>
-    </div>
-  );
-}
