+++ src/app/(root)/page.tsx
@@ -23,17 +23,11 @@ export default function Home() {
       <LinkBtn text="Register Page" path="/register" />
       <LinkBtn text="Lessee Page" path="/home/lesseeView" />
       <LinkBtn text="Admin Page" path="/admin" />
-      <LinkBtn text="LifeStyle Page" path="/lifestyle" />
+      <LinkBtn text="Lifestyle Tag Page" path="/lifestyle" />
       <LinkBtn text="Verification Page" path="/verification" />
       <LinkBtn text="Lessor Rent Income Page" path="/lessorIncome" />
-      <LinkBtn
-        text="Lessor Notification Page"
-        path="/notification/lessorNoti"
-      />
-      <LinkBtn
-        text="Lessee Notification Page"
-        path="/notification/lesseeNoti"
-      />
+      <LinkBtn text="Lessor Notification Page" path="lessorNotification" />
+      <LinkBtn text="Lessee Notification Page" path="/lesseeNotification" />
       <LinkBtn text="Lessor Property Requests" path="/lessorRequest" />
     </div>
   );
