--- src/app/admin/adminHeader.tsx
@@ -1,7 +1,0 @@
-export default function AdminHeader() {
-  return (
-    <div className="text-center">
-      <p className="text-4xl font-bold text-gray-800">Admin View</p>
-    </div>
-  );
-}
