+++ src/app/admin/page.tsx
@@ -1,10 +1,9 @@
-import AdminHeader from '@/app/admin/adminHeader';
-import AdminView from '@/app/admin/adminView';
+import AdminView from '@/components/admin-page/adminView';
 
-export default function AdminPage() {
+export default function Page() {
   return (
     <div className="bg-gray-100 min-h-screen flex flex-col items-center py-10 px-4">
-      <AdminHeader />
+      <p className="text-4xl font-bold text-gray-800">Admin View</p>
       <AdminView />
     </div>
   );
