--- src/components/lessorIncome-page/lessorDashHeader.tsx
@@ -1,7 +1,0 @@
-export default function LessorDashHeader() {
-  return (
-    <div className="text-center">
-      <p className="text-4xl font-bold text-gray-800">Rent Income Dashboard</p>
-    </div>
-  );
-}
