+++ src/components/setting-page/UpdateInformationForm.tsx
@@ -25,7 +25,7 @@ import { useEffect } from 'react';
 import { CircleCheckBig } from 'lucide-react';
 import { useSession } from 'next-auth/react';
 import { useToast } from '@/hooks/use-toast';
-import { UpdateUserInformation } from '@/action/setting/action';
+import { UpdateUserInformation } from '@/actions/setting/action';
 
 const formSchema = z.object({
   username: z
